pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
authors = [
    { name = "Sergey Knyazkov", email = "serg.r88@gmail.com" }
]
readme = "README.md"
requires-python = "~=3.12.9"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]

[tool.ruff]
output-format = "concise"
line-length = 80

[tool.ruff.lint]
extend-select = [
    "E", # pycodestyle Errors
    "F", # Pyflakes Errors
    "W", # pycodestyle Warnings
    "C90", # mccabe
    "I", # isort
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "PL", # Pylint
    "N", # pep8-naming
]
preview = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"
